requirements.txt
streamlit>=1.33.0,<2.0
google-genai>=1.0.0,<2.0
tavily-python>=0.5.0,<1.0
PyPDF2>=3.0.0,<4.0
faiss-cpu>=1.7.4,<2.0
numpy>=1.24.0,<3.0
python-dotenv>=1.0.0,<2.0
requests>=2.31.0,<3.0
